# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Veloce Wear SCM Simulation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /student/dashboard:
    get:
      operationId: getStudentDashboard
      tags: [student]
      summary: Get student dashboard data
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /student/modules/{moduleKey}:
    get:
      operationId: getModuleData
      tags: [student]
      summary: Get module data for student
      parameters:
        - name: moduleKey
          in: path
          required: true
          schema:
            type: string
            enum: [M1, M2, M3]
      responses:
        "200":
          description: Module data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleData"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Module locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /student/modules/{moduleKey}/practice:
    post:
      operationId: runPractice
      tags: [student]
      summary: Run a practice simulation
      parameters:
        - name: moduleKey
          in: path
          required: true
          schema:
            type: string
            enum: [M1, M2, M3]
      responses:
        "200":
          description: Practice run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResult"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /student/modules/{moduleKey}/submit:
    post:
      operationId: submitFinal
      tags: [student]
      summary: Submit final for a module
      parameters:
        - name: moduleKey
          in: path
          required: true
          schema:
            type: string
            enum: [M1, M2, M3]
      responses:
        "200":
          description: Final submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResult"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    ModuleInfo:
      type: object
      properties:
        key:
          type: string
        title:
          type: string
        maxScore:
          type: integer
        status:
          type: string
          enum: [Locked, "Not started", "In progress", Submitted]
        window:
          type: string
          enum: [Open, Closed]
        isUnlocked:
          type: boolean
        score:
          type: integer
          nullable: true
        submittedAt:
          type: string
          nullable: true
        practiceRunCount:
          type: integer
      required: [key, title, maxScore, status, window, isUnlocked, practiceRunCount]
    DashboardData:
      type: object
      properties:
        modules:
          type: array
          items:
            $ref: "#/components/schemas/ModuleInfo"
        totalScore:
          type: integer
        maxTotal:
          type: integer
        totalPracticeRuns:
          type: integer
        userName:
          type: string
      required: [modules, totalScore, maxTotal, totalPracticeRuns, userName]
    SimulationRun:
      type: object
      properties:
        runNumber:
          type: integer
        score:
          type: integer
        isFinal:
          type: boolean
        createdAt:
          type: string
      required: [runNumber, score, isFinal, createdAt]
    ModuleData:
      type: object
      properties:
        moduleKey:
          type: string
        isSubmitted:
          type: boolean
        isUnlocked:
          type: boolean
        finalSubmission:
          nullable: true
          type: object
          properties:
            score:
              type: integer
            maxScore:
              type: integer
            submittedAt:
              type: string
        recentRuns:
          type: array
          items:
            $ref: "#/components/schemas/SimulationRun"
        practiceCount:
          type: integer
      required: [moduleKey, isSubmitted, isUnlocked, recentRuns, practiceCount]
    RunResult:
      type: object
      properties:
        runNumber:
          type: integer
        score:
          type: integer
        isFinal:
          type: boolean
        message:
          type: string
      required: [runNumber, score, isFinal, message]
...