# lib/api-zod/src/generated/types/dashboardData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Veloce Wear SCM Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { ModuleInfo } from "./moduleInfo";

export interface DashboardData {
  modules: ModuleInfo[];
  totalScore: number;
  maxTotal: number;
  totalPracticeRuns: number;
  userName: string;
}