# lib/api-zod/src/generated/types/moduleInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Veloce Wear SCM Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { ModuleInfoStatus } from "./moduleInfoStatus";
import type { ModuleInfoWindow } from "./moduleInfoWindow";

export interface ModuleInfo {
  key: string;
  title: string;
  maxScore: number;
  status: ModuleInfoStatus;
  window: ModuleInfoWindow;
  isUnlocked: boolean;
  score?: number | null;
  submittedAt?: string | null;
  practiceRunCount: number;
}